test_utils.js - move parseFilter into separate file
